Function: Calculate Agreement Part 2
In this part of the code, I am looking for a way to determine the neighbours that are next to an element.

I have analysed 4 cases, which consider the elements around the edges which may not have neighbours above, below, to the left or to the right of them. This is seem in the if statements. If the condition for each input is met, the neighbour is appended to the initially empty list, neighbours.

I changed popuation_shape to m,n so that I could talk about the rows and columns seperately.