Function: Calculate Agreement Part 6
I added else statements in the code  to show the handling for edge cases where the current element is located at the border of the matrix. Basically, they indicate that if the current element is at the top row or leftmost column, the code should consider the element from the opposite border (bottom row or rightmost column) as its neighbour. This wrapping-around behaviour ensures that the neighbouring elements are, hopefully, correctly identified even at the edges of the matrix.