Function: ising_main Part 2
I had to change the file name, which is why it has replaced all the mode. When testing the graph, a plot was appearing for every loop, which was being caused by the plot function in Network, so I modified both ising_main and plot to just update with the new data.